refactor(type): type index 추가
