+++ types/createDocsType.interface.ts
@@ -1,4 +1,4 @@
-export interface CreateDocsType {
+export default interface CreateDocsType {
   title: string;
   enroll: number;
   docsType: string;
