+++ types/history.interface.ts
@@ -1,4 +1,4 @@
-export interface HistoryType {
+export default interface HistoryType {
   thisVersionCreatedAt: Date;
   userId: number;
   nickName: string;
