+++ types/index.ts
@@ -1,0 +1,12 @@
+export type { default as CoinRankType } from "./coinRank.interface";
+export type { default as ContributeType } from "./contribute.interface";
+export type { default as ContributeDocsType } from "./contributeDocs.interface";
+export type { default as CreateDocsType } from "./createDocsType.interface";
+export type { default as DocsItemType } from "./docsItem.interface";
+export type { default as DocsListType } from "./docsList.interface";
+export type { default as DocsListItemType } from "./docsListItem.interface";
+export type { default as HistoryType } from "./history.interface";
+export type { default as ModalState } from "./modal.interface";
+export type { default as StyleVariantsType } from "./styleVariants.type";
+export type { default as TradeType } from "./trade.interface";
+export type { default as UserType } from "./user.interface";
